Creating domain entities for bill types

+++ Bills/Bills.Domain/Entities/Bill.cs
@@ -4,7 +4,7 @@ namespace Bills.Domain
 {
     public abstract class Bill
     {
-        public int TotalToPay { get; private set; }
+        public decimal TotalToPay { get; private set; }
         public Document RecipientIdentification { get; private set; }
         public Address RecipientAddress { get; private set; }
     }
